0608
워프  창 컨트롤

+++ Assets/Scripts/Bullets.cs
@@ -4,8 +4,6 @@ using UnityEngine;
 
 public class Bullets : MonoBehaviour
 {
-    public int damage;
+    public float damage;
 
-
-    
 }
